Insanity. Add object loader; mess up code.
I haven't got a lot of C++ knowledge, and to do this I had to
stretch what knowledge I *have* got to the very limits -- this
included wandering off to StackOverflow a number of times to get
solutions to simple questions (or questions that ought to be simple)
like 'how do I split a string', and 'how do I trim a string'.

The result is a pile of ugly, monstrous code floating around in the
top level scope of the file. This was a first attempt to 'get it
working', and now comes the cleanup phase. I'll be pulling things
into the Object class (and maybe make an 'Objects') class as a
convenience to hold the vector of them), but who knows.

Forays into C++, you are confusing and annoying, and expose me for
the incompetent impostofraud I really am. :)